% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_utils.R
\name{add_es}
\alias{add_es}
\title{Add metaMA effectsize values to top tables.
Used internally by \code{setup_combo_data} and \code{\link[crossmeta]{es_meta}}
to add moderated unbiased standardised effect sizes (dprimes) to top tables
from differential expression analysis.}
\usage{
add_es(diff_exprs, cols = c("dprime", "vardprime"))
}
\arguments{
\item{diff_exprs}{Result from call to \code{\link[crossmeta]{diff_expr}}.}

\item{cols}{Columns from \code{\link[metaMA]{effectsize}} result to add to
top tables.}
}
\value{
diff_exprs with specified columns added to top_tables for each contrast.
}
\description{
Add metaMA effectsize values to top tables.
Used internally by \code{setup_combo_data} and \code{\link[crossmeta]{es_meta}}
to add moderated unbiased standardised effect sizes (dprimes) to top tables
from differential expression analysis.
}
\examples{
library(crossmeta)
library(lydata)

# location of raw data
data_dir <- system.file("extdata", package = "lydata")

# load previous analysis for eset
anal <- load_diff("GSE9601", data_dir)

# add dprime and vardprime to top tables
anal <- add_es(anal)
}
\seealso{
\link[crossmeta]{diff_expr}, \link[crossmeta]{es_meta}.
}

