% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_utils.R
\name{get_dprimes}
\alias{get_dprimes}
\title{Extract unbiased effect sizes from meta-analysis by ES.GeneMeta.}
\usage{
get_dprimes(es2)
}
\arguments{
\item{es2}{ES.GeneMeta.res object, result of of call to \code{ES.GeneMeta}}
}
\value{
List containing:
  \item{meta}{Named numeric vector with overall mean effect size for each gene
     from meta analysis.}
  \item{contrasts}{List of named numeric vectors (one per contrast) with
     unbiased effect sizes for each gene from meta analysis.}
}
\description{
Function extracts MUvals (overall mean effect size) and Effect_Ex (unbiased
effect sizes from each contrast).
}
\details{
Result used to query connectivity map drugs.
}
\seealso{
\link{\code{ES.GeneMeta}}.
}

