% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_utils.R
\name{get_drug_genes}
\alias{get_drug_genes}
\title{Get list of top differentially regulated genes for drugs.}
\usage{
get_drug_genes(drug_info, query_genes, drug_n = nrow(drug_info), es = TRUE)
}
\arguments{
\item{drug_info}{Matrix of differential expression values or gene rankings
(must set es=FALSE) for drugs. Gene rankings should be in increasing order
(e.g. 1 is the most up or down regulated gene). Rows are genes, columns are
drugs.}

\item{query_genes}{Named numeric vector of differentual expression values for
query genes. Usually 'meta' slot of \code{get_dprimes} result.}

\item{drug_n}{Integer specifying how many of the top DE genes should be
used from drug_genes. If there are less than query_n genes in common
between query and drug signatures, the number of common genes will be used.}

\item{es}{Does drug_info contain effect size data? If FALSE, gene ranking
data is assumed.}
}
\value{
List of lists (one per drug in drug_info) with slots:
  \item{up}{Named vector of genes up-regulated by drug.}
  \item{up}{Named vector of genes down-regulated by drug.}
}
\description{
Get list of top differentially regulated genes for drugs.
}
\seealso{
get_top_drugs
}

