% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_utils.R
\name{get_top_combos}
\alias{get_top_combos}
\title{Title}
\usage{
get_top_combos(query_genes, query_n = length(query_genes), drug_info = NULL,
  drug_n = NULL)
}
\arguments{
\item{drug_n}{}
}
\description{
Title
}

