% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_utils.R
\name{get_updn}
\alias{get_updn}
\title{Get number of top DE genes.}
\usage{
get_updn(genes, genes_n)
}
\arguments{
\item{genes}{Named numeric vector of gene effect sizes.}

\item{genes_n}{Integer specifying the desired number of the top DE genes.}
}
\value{
List with slots 'up' and 'dn' each containing character vectors
  of up- and down-regulated genes.
}
\description{
Function returns at most the requested number of genes, attempting to devide
equally between the top up and down regulated genes.
}
\details{
If fewer than half of requested genes are available among up (or down)
regulated genes, then more genes will be taken from down (or up) regulated
genes.
}
\seealso{
\code{\link{setup_query_genes}}, \code{\link{get_drug_genes}}
}

