% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_combos.R
\name{query_combos}
\alias{query_combos}
\title{Title}
\usage{
query_combos(query_genes, db_dir, query_n = length(query_genes),
  ncores = parallel::detectCores(), step = NULL)
}
\arguments{
\item{drug_n}{}
}
\description{
Title
}

