% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_combos.R
\name{range_query_combos}
\alias{range_query_combos}
\title{Title}
\usage{
range_query_combos(query_genes, db_dir, ncores = parallel::detectCores(),
  step = 100)
}
\arguments{
\item{ncores}{}
}
\description{
Title
}

