% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_utils.R
\name{setup_query_genes}
\alias{setup_query_genes}
\title{Get top DE genes for query.}
\usage{
setup_query_genes(query_genes, query_n, drug_info)
}
\arguments{
\item{query_genes}{Named numeric vector of differentual expression values for
query genes. Usually 'meta' slot of \code{get_dprimes} result.}

\item{query_n}{Integer specifying how many of the top DE gene should be used
from query_genes. If there are less than query_n genes in common between
query and drug signatures, the number of common genes will be used.}

\item{drug_info}{Matrix of differential expression values or gene rankings
(must set es=FALSE) for drugs. Gene rankings should be in increasing order
(e.g. 1 is the most up or down regulated gene). Rows are genes, columns are
drugs.}
}
\value{
List with slots 'up' and 'dn' each containing captalized character
  vectors of up- and down-regulated genes.
}
\description{
Get top DE genes for query.
}

