% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_drugs.R
\name{sum_rowcolCumsum}
\alias{sum_rowcolCumsum}
\title{Sum of cumulative sum computed over rows then columns of matrix.}
\usage{
sum_rowcolCumsum(x, i, j)
}
\arguments{
\item{x}{Numeric vector of non-zero values of matrix.}

\item{i}{Integer vector of row indices of x.}

\item{j}{Integer vector of column indices of x.}
}
\value{
Numeric value equal to the sum of the cumulative sum computed over
   rows then columns of a matrix.
}
\description{
Equivalent to computing the cumulative sum of a matrix over rows, then
over columns, then suming every value (though much more fast and memory
efficient).
}
\examples{
x <- c(1, 1, 1, -1) # non-zero values of matrix
i <- c(1, 2, 3, 4)  # row indices of x
j <- c(4, 1, 3, 2)  # col indices of x

sum_rowcolCumsum(x, i, j)
}

