% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_utils.R
\name{get_drug_genes}
\alias{get_drug_genes}
\title{Get list of top differentially regulated genes for drugs.}
\usage{
get_drug_genes(drug_info, query_genes, drug_n = nrow(drug_info), es = TRUE)
}
\value{
List of lists (one per drug in drug_info) with slots:
  \item{up}{Named vector of genes up-regulated by drug.}
  \item{up}{Named vector of genes down-regulated by drug.}
}
\description{
Get list of top differentially regulated genes for drugs.
}
\seealso{
get_top_drugs
}

