% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_combos.R
\name{predict_combos}
\alias{predict_combos}
\title{Get predicted drug combination signatures.}
\usage{
predict_combos(include, exclude = NULL, dat = NULL)
}
\arguments{
\item{with}{Character vector of drug names to combine cmap drugs with.}
}
\value{
drug_info Matrix of differential expression values for drug
  combinations. Rows are genes, columns are drugs.
}
\description{
Predicts drug combination signatures.
}
\details{
Drug combinations may more closely mimic or reverse the query signature than
individual drugs. The drug combinations that most closely mimic or reverse
a query signature can usually be determined by querying against drug
combination signatures predicted with the top few single drugs.
}
\examples{
# this returns NULL:
combos_info <- predict_combos("cmap_drug")


# predicted expression values for combined treatment with sirolimus
# and all other cmap drugs:
# combos_info <- predict_combos("sirolimus")
}
\seealso{
\code{\link{query_drugs}} to determine overlap between query and
   predicted drug combination signatures.
}

