% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_combos.R
\name{query_combos}
\alias{query_combos}
\title{Get overlap between query and drug combination signatures.}
\usage{
query_combos(query_genes, db_dir, query_n = length(query_genes),
  ncores = parallel::detectCores(), step = NULL)
}
\arguments{
\item{query_genes}{Named numeric vector of differentual expression values for
query genes. Usually 'meta' slot of \code{get_dprimes} result.}

\item{db_dir}{String specifying full path to drug_combos.sqlite database.}

\item{query_n}{Integer specifying how many of the top DE gene should be used
from query_genes. If there are less than query_n genes in common between
query and drug signatures, the number of common genes will be used.
Parameter ignored if \code{step} not \code{NULL}.}

\item{ncores}{Number of cores to use for parallel queries. Default is all
available.}

\item{step}{Integer specifying step size for query range. Queries start
at 100 genes, and step up by specified step size up to the number of
common genes.}
}
\value{
List of data.frames (one per query size) each with columns:
  \item{drug_n}{Number of drug combination genes used.}
  \item{query_n}{Number of query genes used.}
  \item{overlap}{Number of genes that change in the same direction in drug
     combination and query signatures.}
  \item{cross}{Number of genes that change in the opposite direction in drug
     combination and query signatures.}
  \item{net}{Difference between overlap and cross.}
}
\description{
Determines the number of genes that change in the same direction between
query and predicted drug combination signatures.
}
\details{
All 856086 combinations of the 1309 drugs from the Connectivity Map build 02
release are compared to the query signature. Drug combinations with the
largest positive and negative net overlap are predicted to, respectively,
mimic and reverse the query signature. Drug combinations may more closely
mimic or reverse the query signature than individual drugs.
}
\examples{
library(crossmeta)
library(lydata)

db_dir <- "path/to/drug_combos.sqlite"

# location of data
data_dir <- system.file("extdata", package = "lydata")

# gather GSE names
gse_names  <- c("GSE9601", "GSE15069", "GSE50841", "GSE34817", "GSE29689")

# load previous analysis
anals <- load_diff(gse_names, data_dir)

# perform meta-analysis
es <- es_meta(anals)

# get query signature
dprimes <- get_dprimes(es)

# query drug combination database
# top_combos <- query_combos(dprimes$meta, db_dir)
}
\seealso{
\code{\link{make_drug_combos}} to create drug combination database.

Other query.functions: \code{\link{query_drugs}},
  \code{\link{range_query_combos}},
  \code{\link{range_query_drugs}}
}

