% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_drugs.R
\name{query_drugs}
\alias{query_drugs}
\title{Get overlap between query and drug signatures.}
\usage{
query_drugs(query_genes, drug_info = NULL, sorted = TRUE)
}
\arguments{
\item{query_genes}{Named numeric vector of differentual expression values for
query genes. Usually 'meta' slot of \code{get_dprimes} result.}

\item{drug_info}{Matrix of differential expression values for drugs or drug
combinations. Rows are genes, columns are drugs.}

\item{sorted}{Would you like the results sorted in decreasing order of overlap?
Default is TRUE.}
}
\value{
Vector of numeric values between 1 and -1 indicating extent of overlap
  between query and drug signatures (see description).
}
\description{
Determines the volume under the surface formed by plotting net overlap (z)
as a function of number of drug and query genes (x and y).
}
\details{
Drugs with the largest positive and negative net overlap are predicted to,
respectively, mimic and reverse the query signature. A value of 1 would indicate
that all drug and query genes are regulated in the same direction and have the
same order when sorted by absolute changes in differential expression. A value
of -1 would indicate that all drug and query genes are regulated in the
opposite direction and have the same order when sorted by absolute changes
in differential expression.
}
\examples{

# create drug signatures
genes <- paste("GENE", 1:1000, sep = "_")
set.seed(0)

drug_info <- data.frame(row.names = genes,
                        drug1  = rnorm(1000, sd = 2),
                        drug2  = rnorm(1000, sd = 2),
                        drug3  = rnorm(1000, sd = 2))

# query signature is drug3
query_sig <- drug_info$drug3
names(query_sig) <- genes

res <- query_drugs(query_sig, as.matrix(drug_info))
}
\seealso{
\code{\link{query_combos}} to get overlap between query and
  predicted drug combination signatures.
}

