% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_drugs.R
\name{query_drugs}
\alias{query_drugs}
\title{Get overlap between query and drug signatures.}
\usage{
query_drugs(query_genes, drug_info = NULL)
}
\arguments{
\item{query_genes}{Named numeric vector of differentual expression values for
query genes. Usually 'meta' slot of \code{get_dprimes} result.}

\item{drug_info}{Matrix of differential expression values for drugs or drug
combinations. Rows are genes, columns are drugs.}
}
\value{
Vector of numeric values between 1 and -1 indicating extent of overlap
  between query and drug signatures (see details).
}
\description{
Determines the volume under the surface formed by plotting net overlaps (z)
as a function of number of drug and query genes (x and y).
}
\details{
Drugs with the largest positive and negative net overlap are predicted to,
respectively, mimic and reverse the query signature. A value of 1 would indicate
that all drug and query genes are regulated in the same direction and with
the same rankings within their signatures. A value of -1 would indicate
that all drug and query genes are regulated in the opposite direction and with
the same rankings within their signatures.
}
\examples{

# create drug signatures
genes <- paste("GENE", 1:1000, sep = "_")
set.seed(0)

drug_info <- data.frame(row.names = genes,
                        drug1  = rnorm(1000, sd = 2),
                        drug2  = rnorm(1000, sd = 2),
                        drug3  = rnorm(1000, sd = 2))

# query signature is drug3
query_sig <- drug_info$drug3
names(query_sig) <- genes

res <- query_drugs(query_sig, as.matrix(drug_info))
}
\seealso{
\code{\link{predict_combos}} to get predicted drug combination
  signatures (can be passed to \code{drug_info}).
}

