% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_drugs.R
\name{query_drugs}
\alias{query_drugs}
\title{Get overlap between query and drug signatures.}
\usage{
query_drugs(query_genes, drug_info, query_n = length(query_genes),
  drug_n = nrow(drug_info), step = NULL, es = TRUE)
}
\arguments{
\item{query_genes}{Named numeric vector of differentual expression values for
query genes. Usually 'meta' slot of \code{get_dprimes} result.}

\item{drug_info}{Matrix of differential expression values or gene rankings
(must set es=FALSE) for drugs. Gene rankings should be in increasing order
(e.g. 1 is the most up or down regulated gene). Rows are genes, columns are
drugs.}

\item{query_n}{Integer specifying how many of the top DE gene should be used
from query_genes. If there are less than query_n genes in common between
query and drug signatures, the number of common genes will be used.}

\item{drug_n}{Integer specifying how many of the top DE genes should be
used from drug_genes. If there are less than query_n genes in common
between query and drug signatures, the number of common genes will be used.}

\item{es}{Does drug_info contain effect size data? If FALSE, gene ranking
data is assumed.}

\item{drug_genes}{List of lists (one per drug) with slots 'up' and 'dn' each
containing a character vector with the names of up- and down-regulated
genes. If not supplied, created by \code{get_drug_genes}. Useful only if
multiple calls to \code{get_top_drugs} needed (e.g. in \code{test_ma}).}
}
\value{
data.frame with columns:
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
}
\description{
Get overlap between query and drug signatures.
}

