% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_utils.R
\name{setup_query_genes}
\alias{setup_query_genes}
\title{Get top DE genes for query.}
\usage{
setup_query_genes(query_genes, query_n, drug_info)
}
\value{
List with slots 'up' and 'dn' each containing captalized character
  vectors of up- and down-regulated genes.
}
\description{
Get top DE genes for query.
}

